# Add support for DMU_OTN_* types in dbufstat.py