<!--- Provide a general summary of your changes in the Title above -->

<!---
Documentation on ZFS Buildbot options can be found at
https://github.com/zfsonlinux/zfs/wiki/Buildbot-Options
-->

### Description
<!--- Describe your changes in detail -->
This enable us to view DMU_OTN_* types with dbufstat.py.

### Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->
I was investigating the ARC meta leak/accounting issue when i noticed some dbuf type values were not translated to their type names (`196` instead of `DMU_OTN_ZAP_METADATA`).

Additional notes:

1. I've added the new types as hardcoded values to avoid the amount of work required to dynamically calculate them at runtime (every time we execute `get_typestring()`)
2. `DMU_OTN_*_ENC_(META)DATA` values are included for completeness: they will probably never be used, AFAIK (meta)data is not encrypted in memory.

### How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->
<!--- If your change is a performance enhancement, please provide benchmarks here. -->
Manual testing on debian box:

Before:
```
           pool  objset      object                        dtype  cached  
       testpool       0          59                          196     512  
       testpool       0          58                          196     512  
       testpool       0          61                 DMU_OT_BPOBJ     16K  
```

After:
```
           pool  objset      object                        dtype  cached  
       testpool       0          59         DMU_OTN_ZAP_METADATA     512  
       testpool       0          58         DMU_OTN_ZAP_METADATA     512  
       testpool       0          61                 DMU_OT_BPOBJ     16K  
```

### Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Performance enhancement (non-breaking change which improves efficiency)
- [ ] Code cleanup (non-breaking change which makes code smaller or more readable)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Documentation (a change to man pages or other documentation)

### Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the ZFS on Linux code style requirements.
- [ ] I have updated the documentation accordingly.
- [x] I have read the **CONTRIBUTING** document.
- [ ] I have added tests to cover my changes.
- [ ] All new and existing tests passed.
- [x] All commit messages are properly formatted and contain `Signed-off-by`.
- [ ] Change has been approved by a ZFS on Linux member.